Power: Change register width to 64 bits
This changeset sets the width of general-purpose and some of the
other important registers to 64 bits. This is a prerequisite for
enabling the execution of 64-bit, little endian Power executables.
This allows the register operands provided in instructions to be
recognized as 64 bit. Also, following the naming conventions from
Power ISA v3.0B, the PC and NPC registers have beeb renamed to CIA
and NIA respectively.

Change-Id: Ic0a2495fb6c199a4d18ff34d3a57488bb1871a4c