Power: Refactor code and introduce extended opcode fields
This changeset reorders the defined instructions by their category.
The ordering used here is the same as that of the Power ISA manual
which is as follows:
  * Branch facility
  * Fixed-point facility
  * Floating-point facility

Furthermore, the fixed-point instructions are also reordered within
every decode and format block in the same order as that of the ISA
manual which is as follows:
  * Load instructions
  * Store instructions
  * Arithmetic instructions
  * Compare instructions
  * Logical instructions
  * Rotate and shift instructions
  * Move to/from system register instructions

This also introduces separate extended opcode fields for DS, X, XFL,
XFX, XL and XO form instructions to address scenarios where multiple
instructions of different forms share the same primary opcode. The
primary opcode field, currently declared as OPCODE, is also renamed
to PO.

Change-Id: Idd446f0db7d9811594eb15ae002f52ea3f46043a